requirements.txt
python-telegram-bot
schedule
curl_cffi
beautifulsoup4
requests
PyGithub
streamlit
openai
gspread
duckduckgo-search
google-generativeai
plotly
cloudscraper
pyTelegramBotAPI
